detector/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pillow==11.0.0
numpy==2.1.1
python-multipart==0.0.12
# YOLOv8 (실제 모델 사용 시 주석 해제)
# ultralytics==8.3.0
# ONNX Runtime (ONNX 모델 사용 시 주석 해제)
# onnxruntime==1.19.0
# OpenCV (캘리브레이션 사용 시 주석 해제)
# opencv-python==4.10.0.84
# opencv-contrib-python==4.10.0.84
